+++ src/lib/components/Tracklist/TracklistItem.svelte
@@ -1,16 +1,24 @@
 <script lang="ts">
-  import { Button, ListItem } from "m3-svelte";
+  import { Button, Icon, ListItem } from "m3-svelte";
 
   import { formatSeconds } from "$lib/player/format";
   import type { AudioItem } from "@toil/vk-audio/types/client/section";
+  import iconFavoriteOutlineRounded from "@ktibow/iconset-material-symbols/favorite-outline-rounded";
+  import iconHeartBrokenRounded from "@ktibow/iconset-material-symbols/heart-broken-rounded";
 
   export type Props = {
     audio: AudioItem;
     isSelected?: boolean;
     onclick?: (audio: AudioItem) => Promise<void> | void;
+    onlikeToggle?: (audio: AudioItem, isLiked: boolean) => Promise<void> | void;
   };
 
-  let { audio, isSelected = false, onclick = () => void 0 }: Props = $props();
+  let {
+    audio,
+    isSelected = false,
+    onclick = () => void 0,
+    onlikeToggle = () => void 0,
+  }: Props = $props();
 </script>
 
 <li class="tracklist-item" class:selected={isSelected}>
@@ -28,9 +36,28 @@
       />
     </div>
     <ListItem headline={audio.title} overline={audio.artist}></ListItem>
-    <p class="tracklist-item__duration">
-      {formatSeconds(audio.duration)}
-    </p>
+    <div class="tracklist-item__end">
+      <p class="tracklist-item__actions">
+        <Button
+          variant="text"
+          iconType="full"
+          size="s"
+          onclick={async (event: Event) => {
+            event.stopPropagation();
+            await onlikeToggle(audio, audio.isLiked);
+          }}
+        >
+          <Icon
+            icon={audio.isLiked
+              ? iconHeartBrokenRounded
+              : iconFavoriteOutlineRounded}
+          />
+        </Button>
+      </p>
+      <p class="tracklist-item__duration">
+        {formatSeconds(audio.duration)}
+      </p>
+    </div>
   </Button>
 </li>
 
@@ -39,6 +66,14 @@
     width: 100%;
     justify-content: start;
     text-align: left;
+  }
+
+  .tracklist-item__actions {
+    display: none;
+  }
+
+  .tracklist-item:hover .tracklist-item__actions {
+    display: flex;
   }
 
   .tracklist-item__image-wrapper {
@@ -55,10 +90,16 @@
   }
 
   .tracklist-item__duration {
-    margin-left: auto;
     font-size: 0.75rem;
     font-weight: 500;
     color: var(--m3c-on-surface-variant);
+  }
+
+  .tracklist-item__end {
+    display: flex;
+    gap: 1rem;
+    align-items: center;
+    margin-left: auto;
   }
 
   .tracklist-item.selected .tracklist-item__duration {